fix(security) Add authorization check for account deletion

+++ app/Http/Controllers/Account/AccountDeleteController.php
@@ -1,19 +1,23 @@
-<?php
-
-declare(strict_types=1);
+<?php declare(strict_types=1);
 
 namespace App\Http\Controllers\Account;
 
 use App\Models\Account;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class AccountDeleteController
 {
     public function delete(Request $request, int $id)
     {
-        // TODO add security check, add can('account delete')
+        if (Auth::user()->cannot('account delete')) {
+            return redirect()->back()->with('error', 'You do not have permission to delete accounts.');
+        }
+
         $account = Account::find($id);
-        $account->delete();
+        if ($account) {
+            $account->delete();
+        }
 
         return redirect()->back();
     }
